Re-organize the application un/installation code
1) The appplication un/installation now is Tizen-only.
2) The code which is only used by 'xwalkctl' has moved
   to 'application/tools/tizen'.
3) The unneeded platform abstractions were removed.
