+++ application/common/package/package.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "xwalk/application/common/installer/package.h"
+#include "xwalk/application/common/package/package.h"
 
 #include "base/file_util.h"
 #include "base/files/file_path.h"
@@ -10,8 +10,8 @@
 #include "base/path_service.h"
 #include "third_party/zlib/google/zip.h"
 #include "xwalk/application/common/id_util.h"
-#include "xwalk/application/common/installer/wgt_package.h"
-#include "xwalk/application/common/installer/xpk_package.h"
+#include "xwalk/application/common/package/wgt_package.h"
+#include "xwalk/application/common/package/xpk_package.h"
 
 namespace xwalk {
 namespace application {
