+++ application/common/package/package.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef XWALK_APPLICATION_COMMON_INSTALLER_PACKAGE_H_
-#define XWALK_APPLICATION_COMMON_INSTALLER_PACKAGE_H_
+#ifndef XWALK_APPLICATION_COMMON_PACKAGE_PACKAGE_H_
+#define XWALK_APPLICATION_COMMON_PACKAGE_PACKAGE_H_
 
 #include <string>
 #include <vector>
@@ -60,4 +60,4 @@ class Package {
 }  // namespace application
 }  // namespace xwalk
 
-#endif  // XWALK_APPLICATION_COMMON_INSTALLER_PACKAGE_H_
+#endif  // XWALK_APPLICATION_COMMON_PACKAGE_PACKAGE_H_
