+++ application/common/package/wgt_package.h
@@ -2,15 +2,15 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef XWALK_APPLICATION_COMMON_INSTALLER_WGT_PACKAGE_H_
-#define XWALK_APPLICATION_COMMON_INSTALLER_WGT_PACKAGE_H_
+#ifndef XWALK_APPLICATION_COMMON_PACKAGE_WGT_PACKAGE_H_
+#define XWALK_APPLICATION_COMMON_PACKAGE_WGT_PACKAGE_H_
 
 #include <string>
 #include <vector>
 
 #include "base/files/file_path.h"
 #include "base/memory/scoped_ptr.h"
-#include "xwalk/application/common/installer/package.h"
+#include "xwalk/application/common/package/package.h"
 
 namespace xwalk {
 namespace application {
@@ -24,4 +24,4 @@ class WGTPackage : public Package {
 }  // namespace application
 }  // namespace xwalk
 
-#endif  // XWALK_APPLICATION_COMMON_INSTALLER_WGT_PACKAGE_H_
+#endif  // XWALK_APPLICATION_COMMON_PACKAGE_WGT_PACKAGE_H_
