+++ application/common/package/xpk_package.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "xwalk/application/common/installer/xpk_package.h"
+#include "xwalk/application/common/package/xpk_package.h"
 
 #include "base/file_util.h"
 #include "base/files/scoped_file.h"
