+++ application/common/tizen/configuration/signature_schema.xsd
