+++ application/common/tizen/signature_data.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef XWALK_APPLICATION_COMMON_INSTALLER_SIGNATURE_DATA_H_
-#define XWALK_APPLICATION_COMMON_INSTALLER_SIGNATURE_DATA_H_
+#ifndef XWALK_APPLICATION_COMMON_TIZEN_SIGNATURE_DATA_H_
+#define XWALK_APPLICATION_COMMON_TIZEN_SIGNATURE_DATA_H_
 
 #include <list>
 #include <set>
@@ -117,4 +117,4 @@ class SignatureData {
 }  // namespace application
 }  // namespace xwalk
 
-#endif  // XWALK_APPLICATION_COMMON_INSTALLER_SIGNATURE_DATA_H_
+#endif  // XWALK_APPLICATION_COMMON_TIZEN_SIGNATURE_DATA_H_
