+++ application/common/tizen/signature_parser.h
@@ -2,12 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef XWALK_APPLICATION_COMMON_INSTALLER_SIGNATURE_PARSER_H_
-#define XWALK_APPLICATION_COMMON_INSTALLER_SIGNATURE_PARSER_H_
+#ifndef XWALK_APPLICATION_COMMON_TIZEN_SIGNATURE_PARSER_H_
+#define XWALK_APPLICATION_COMMON_TIZEN_SIGNATURE_PARSER_H_
 
 #include "base/files/file_path.h"
 #include "base/memory/scoped_ptr.h"
-#include "xwalk/application/common/installer/signature_data.h"
+#include "xwalk/application/common/tizen/signature_data.h"
 
 namespace xwalk {
 namespace application {
@@ -24,4 +24,4 @@ class SignatureParser {
 }  // namespace application
 }  // namespace xwalk
 
-#endif  // XWALK_APPLICATION_COMMON_INSTALLER_SIGNATURE_PARSER_H_
+#endif  // XWALK_APPLICATION_COMMON_TIZEN_SIGNATURE_PARSER_H_
