+++ application/common/xwalk_application_common.gypi
@@ -49,19 +49,12 @@
         'permission_policy_manager.h',
         'permission_types.h',
         'signature_types.h',
-
-        'installer/package.h',
-        'installer/package.cc',
-        'installer/package_installer.cc',
-        'installer/package_installer.h',
-        'installer/signature_data.h',
-        'installer/signature_data.cc',
-        'installer/signature_parser.h',
-        'installer/signature_parser.cc',
-        'installer/wgt_package.h',
-        'installer/wgt_package.cc',
-        'installer/xpk_package.cc',
-        'installer/xpk_package.h',
+        'package/package.h',
+        'package/package.cc',
+        'package/wgt_package.h',
+        'package/wgt_package.cc',
+        'package/xpk_package.cc',
+        'package/xpk_package.h',
       ],
       'conditions': [
         ['tizen==1', {
@@ -83,14 +76,15 @@
             'manifest_handlers/tizen_setting_handler.h',
             'manifest_handlers/tizen_splash_screen_handler.cc',
             'manifest_handlers/tizen_splash_screen_handler.h',
-            'installer/package_installer_tizen.cc',
-            'installer/package_installer_tizen.h',
-            'installer/tizen/packageinfo_constants.cc',
-            'installer/tizen/packageinfo_constants.h',
-            'installer/tizen/signature_validator.cc',
-            'installer/tizen/signature_validator.h',
             'tizen/package_path.cc',
             'tizen/package_path.h',
+            'tizen/signature_data.h',
+            'tizen/signature_data.cc',
+            'tizen/signature_parser.h',
+            'tizen/signature_parser.cc',
+            'tizen/signature_validator.cc',
+            'tizen/signature_validator.h',
+
           ],
         }, {
         'sources': [
