+++ xwalk_tests.gypi
@@ -34,7 +34,7 @@
       ],
       'sources': [
         'application/common/application_storage_impl_unittest.cc',
-        'application/common/installer/package_unittest.cc',
+        'application/common/package/package_unittest.cc',
         'application/common/application_unittest.cc',
         'application/common/application_file_util_unittest.cc',
         'application/common/id_util_unittest.cc',
